Split manually vectorized code into separate files.
