+++ src/nnue/layers/input_slice.h
@@ -30,7 +30,7 @@ template <IndexType OutDims, IndexType Offset = 0>
 class InputSlice {
  public:
   // Need to maintain alignment
-  static_assert(Offset % MaxSimdWidth == 0, "");
+  static_assert(Offset % 32 == 0, "");
 
   // Output type
   using OutputType = TransformedFeatureType;
